+++ cdk_framework/EKS/lib/resource_constructs/general_constructs/sample-app-service-construct.ts
@@ -1,6 +1,6 @@
 import { Construct } from 'constructs';
 import { Cluster, FargateCluster } from 'aws-cdk-lib/aws-eks';
-import { NamespaceConstruct } from './namespace-construct';
+import { NamespaceConstruct } from '../specific_constructs/namespace-construct';
 
 export class SampleAppServiceConstruct extends Construct {
     name: string
