+++ cdk_framework/EKS/lib/resource_constructs/specific_constructs/collector-config-map-construct.ts
