+++ cdk_framework/EKS/lib/resource_constructs/specific_constructs/grpc-service-construct.ts
