+++ cdk_framework/EKS/lib/resource_constructs/specific_constructs/pull-mode-sample-app-construct.ts
