+++ cdk_framework/EKS/lib/resource_constructs/specific_constructs/tcp-service-construct.ts
